+++ tests/test_decoder.py
@@ -29,6 +29,26 @@ class BaseTestCase(unittest.TestCase):
             },
         )
 
+    def test_decoder_geojson(self):
+        vector_tile = b'\x1aI\n\x05water\x12\x1a\x08\x01\x12\x06\x00\x00\x01\x01\x02\x02\x18\x03"\x0c\t\x00\x80@\x1a\x00\x01\x02\x00\x00\x02\x0f\x1a\x03foo\x1a\x03baz\x1a\x03uid"\x05\n\x03bar"\x05\n\x03foo"\x02 {(\x80 x\x02'  # noqa
+        self.assertEqual(
+            mapbox_vector_tile.decode(vector_tile, geojson=False),
+            {
+                "water": {
+                    "version": 2,
+                    "extent": 4096,
+                    "features": [
+                        {
+                            "geometry": {"type": "Polygon", "coordinates": [[[0, 0], [0, 1], [1, 1], [1, 0], [0, 0]]]},
+                            "properties": {"foo": "bar", "baz": "foo", "uid": 123},
+                            "id": 1,
+                            "type": 3,
+                        }
+                    ],
+                },
+            },
+        )
+
     def test_decode_polygon_no_cmd_seg_end(self):
         # the binary here was generated without including the
         # CMD_SEG_END after the polygon parts